+++ plugins/parsers/json_v2/parser_test.go
@@ -100,6 +100,14 @@ func TestMultipleConfigs(t *testing.T) {
 	}
 }
 
+func TestParserEmptyConfig(t *testing.T) {
+	plugin := &json_v2.Parser{
+		Configs: []json_v2.Config{},
+	}
+
+	require.ErrorContains(t, plugin.Init(), "no configuration provided")
+}
+
 func BenchmarkParsingSequential(b *testing.B) {
 	inputFilename := filepath.Join("testdata", "benchmark", "input.json")