+++ CHANGELOG.md
@@ -14,6 +14,9 @@
   As a consequence the redunant `logtarget` setting is deprecated, `stderr` is
   used if no `logfile` is provided, otherwise messages are logged to the given
   file. For using the Windows `eventlog` set `logformat = "eventlog"`!
+- This release contains a change in json_v2 parser config parsing -
+  if the config is empty (not define any rules), initialization will fail
+  (see PR [#15844](https://github.com/influxdata/telegraf/pull/15844)).
 
 ## v1.31.3 [2024-08-12]